requirements.txt
# ── 데이터 수집 ───────────────────────────────────────
pykrx>=1.2.4
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
yfinance>=0.2.40

# ── 기술적 지표 ───────────────────────────────────────
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.26.0

# ── AI / Gemini ───────────────────────────────────────
google-genai>=1.0.0

# ── MCP ──────────────────────────────────────────────
mcp>=1.0.0

# ── Slack ─────────────────────────────────────────────
slack-bolt>=1.18.0
slack-sdk>=3.27.0

# ── 스케줄러 ──────────────────────────────────────────
apscheduler>=3.10.0

# ── 유틸리티 ──────────────────────────────────────────
python-dotenv>=1.0.0
